+++ mods/src/config.h
@@ -146,13 +146,14 @@ public:
   static constexpr size_t MaxToastStates = 64;
 
   bool enabled                  = false;
-  bool incoming_attack_player   = false;
-  bool incoming_attack_hostile  = false;
-  bool fleet_arrived_in_system  = false;
-  bool fleet_started_mining     = false;
-  bool fleet_node_depleted      = false;
-  bool fleet_docked             = false;
-  bool fleet_repair_complete    = false;
+  bool incoming_attack_player       = false;
+  bool incoming_attack_hostile      = false;
+  bool fleet_arrived_in_system      = false;
+  bool fleet_arrived_at_destination = false;
+  bool fleet_started_mining         = false;
+  bool fleet_node_depleted          = false;
+  bool fleet_docked                 = false;
+  bool fleet_repair_complete        = false;
 
   [[nodiscard]] bool AnyIncomingAttackEnabled() const
   {